Confirmed Token Reading, Identifiers, Variable Types
Took me a while, but I confirmed that the Lexicial Analyzer appears to be working. I fed it an input string with a variable declaration, values, and a if statement using that newly created identifier and it managed to keep track properly!

There is an issue with numbers greater than 2 characters that they will seperate out, but that seems easy to fix

Someone should create a test sheet to ensure it works for every operator but I feel confident we found it!