Coupled Object Time!
Added all Coupling Object Abstract and Implementables. Since we only have 3 required values ( String Int Boolean ), that makes it a lot easier for us when we group statements  we can make sure they return a value properly!

To do is to actually make the coupling work, which will be dependant on if we are required to move to spacing we might need to fix that.

Also to create all the object we'll need d_d which isn't going to be fun!