Big Fixes with IntParenthesis and Int Add/Multiply
There is an outstanding bug where there are situations where + can resolve before * if the * is waiting on another operator. A way to solve this is to ensure that a operator with greater priority doesn't exist unresolved to the left or right of the unary operator. While more expensive would ensure proper order